overhaul: Show Backpack Preview Feature - add: personal compactors are now colored according to disabled/enabled - add: caching tooltip data for renders as an optimization - add: Show Preview for Builder's Tools FeatureSetting - fix: freeze mode tooltips broken with NEU exist - fix: item tooltips are broken on Freeze Backpack Preview button - fix: white display name on containers other than backpacks - fix: compact mode top spacing isn't equal to bottom

+++ src/main/java/codes/biscuit/skyblockaddons/features/backpacks/BackpackColor.java
@@ -33,15 +33,15 @@ public enum BackpackColor {
     }
 
     public float getR() {
-        return (float)r/255;
+        return r/255F;
     }
 
     public float getG() {
-        return (float)g/255;
+        return g/255F;
     }
 
     public float getB() {
-        return (float)b/255;
+        return b/255F;
     }
 
     private static final Set<BackpackColor> darkColors = EnumSet.of(BackpackColor.BLACK, BackpackColor.PURPLE,
